improve saved cached reponses

+++ save_cached_responses.py
@@ -19,7 +19,7 @@ from pub import Dandelion
 from util import elapsed
 from util import safe_commit
 from util import TooManyRequestsException
-from query_stopwords import stop_words
+from entity import stop_words
 from search import CachedEntityResponse
 from views import get_search_query
 
@@ -51,15 +51,18 @@ def cache_api_response(my_saved_object):
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Run stuff.")
-    parser.add_argument('--since', nargs="?", type=str, default=False, help="just add twitter")
-    parsed = parser.parse_args()
+    parser.add_argument('--since', nargs="?", type=str, default=False, help="date since to dirty cache")
+    parser.add_argument('--entity', nargs="?", type=str, default=False, help="entity to refresh")
+    parsed_args = parser.parse_args()
+    parsed_vars = vars(parsed_args)
 
     start = time()
 
     before_date = datetime.datetime.utcnow()
 
     if __name__ == '__main__':
-        while True:
+        something_to_do = True
+        while something_to_do:
             # query = u"""select * from search_autocomplete_dandelion_mv
             #     where entity_title not in (select entity_title from cached_entity_response)
             #     order by sum_num_events desc
@@ -82,29 +85,36 @@ if __name__ == "__main__":
             #     my_saved_objects.append(my_saved_object)
 
 
-            single_entity_title = None
+            before_date = parsed_vars.get("before", None)
+            debug = parsed_vars.get("debug", None)
+            single_entity_title = parsed_vars.get("entity", None)
             # single_entity_title =  'Gluten-free diet'
+
+            my_saved_objects = []
             if single_entity_title:
-                my_saved_objects = CachedEntityResponse.query.filter(CachedEntityResponse.entity_title == single_entity_title).\
+                my_saved_objects = CachedEntityResponse.query.filter(CachedEntityResponse.entity_title.ilike(single_entity_title)).\
                     order_by(CachedEntityResponse.collected.asc()).limit(25).all()
-            else:
+                something_to_do = False  # only do this once
+
+            elif before_date:
                 my_saved_objects = CachedEntityResponse.query.\
                     filter(or_(CachedEntityResponse.collected==None, CachedEntityResponse.collected < before_date)).\
                     order_by(CachedEntityResponse.collected.asc()).limit(25).all()
 
-            use_threads = True  # useful to turn off pooling to help debugging
-            my_thread_pool = ThreadPool(20)
-            if use_threads:
-                results = my_thread_pool.imap_unordered(cache_api_response, my_saved_objects)
-                my_thread_pool.close()
-                my_thread_pool.join()
+            if my_saved_objects:
+                use_threads = not debug  # useful to turn off pooling to help debugging
+                if use_threads:
+                    my_thread_pool = ThreadPool(20)
+                    results = my_thread_pool.imap_unordered(cache_api_response, my_saved_objects)
+                    my_thread_pool.close()
+                    my_thread_pool.join()
+                    my_thread_pool.terminate()
+                else:
+                    for my_saved_object in my_saved_objects:
+                        cache_api_response(my_saved_object)
             else:
-                for my_saved_object in my_saved_objects:
-                    cache_api_response(my_saved_object)
-
-            my_thread_pool.terminate()
-
-            # print results
+                print "nothing to do"
+                something_to_do = False
 
             db.session.remove()
             # print "finished update in {}sec".format(elapsed(start))
